+++ src/model.rs
@@ -67,7 +67,7 @@ impl<M: ProbabilisticModel> ProbabilisticModel for OrigamiModel<M> {
     // #[inline(always)]
     fn expansion_unnormalized_ll(&self, prod: &Node, expr: &PartialExpr, hole: &Hole) -> NotNan<f32> {
         // if this is not the very first expansion, we forbid the fix1() operator
-        if expr.expr.len() != 0 {
+        if !expr.expr.is_empty() {
             if let Node::Prim(p) = prod  {
                 if *p == self.fix1 {    
                     return NotNan::new(f32::NEG_INFINITY).unwrap();
@@ -76,7 +76,7 @@ impl<M: ProbabilisticModel> ProbabilisticModel for OrigamiModel<M> {
         }
 
         // if this is the very first expansion, we require it to be the fix1() operator
-        if expr.expr.len() == 0  {
+        if expr.expr.is_empty()  {
             if let Node::Prim(p) = prod  {
                 if *p != self.fix1 {    
                     return NotNan::new(f32::NEG_INFINITY).unwrap();
@@ -105,8 +105,8 @@ impl<M: ProbabilisticModel> ProbabilisticModel for OrigamiModel<M> {
 
         // we forbid the use of the very outermost argument if we used a fix1 at the top level
         if let Node::Var(i,-1) = prod {
-            if *i+1 == hole.env.len() as i32 {
-                if expr.expr.len() != 0 {
+            if *i+1 == hole.env.len() as i32
+                && !expr.expr.is_empty() {
                     if let Node::App(f,_) = expr.expr[0] {
                         if let Node::App(f,_) = expr.expr[f] {
                             if expr.expr[f] == Node::Prim(self.fix1.clone()) {
@@ -115,10 +115,7 @@ impl<M: ProbabilisticModel> ProbabilisticModel for OrigamiModel<M> {
                         }
                     }
                 }
-            }
         }
-
-
         // default
         self.model.expansion_unnormalized_ll(prod, expr, hole)
     }
