+++ src/task.rs
@@ -34,7 +34,7 @@ pub fn parse_tasks<D: Domain>(path: &dyn AsRef<Path>, dsl: &DSL<D>) -> Vec<Task<
     let json: serde_json::Value = from_reader(File::open(path).expect("file not found")).expect("json deserializing error");
     let tasks: Vec<Task<D>> = json.as_array().unwrap().iter().map(|task| {
         Task::new(
-            task["name"].as_str().unwrap().into(),
+            task["name"].as_str().unwrap(),
             task["tp"].as_str().unwrap().parse().unwrap(),
             task["ios"].as_array().unwrap().iter().map(|io| {
                 let inputs: Vec<String> = io.as_array().unwrap()[0].as_array().unwrap().iter().map(|i| i.as_str().unwrap().to_string()).collect();
