+++ src/top_down.rs
@@ -202,7 +202,7 @@ impl Expansion {
         expr.ll = self.ll;
         expr.prev_prod = Some(prod.clone());
 
-        let mut to_expand: Option<Idx> = hole.parent.clone();
+        let mut to_expand: Option<Idx> = hole.parent;
 
         // add a new hole for each arg, along with any apps and lams
         for arg_idx in (0..prod_tp.arity(&expr.ctx)).rev() {
@@ -230,7 +230,7 @@ impl Expansion {
             }
 
             // the hole type is the return type of the arg (bc all lambdas were autofilled)
-            let new_hole_tp = arg_tp.return_type(&expr.ctx).clone();
+            let new_hole_tp = arg_tp.return_type(&expr.ctx);
             expr.holes.push(Hole::new(new_hole_tp, new_hole_env, Some(idx)))
         }
 
@@ -261,7 +261,7 @@ pub fn add_expansions(state: &mut ThreadState, prods: &[Prod], model: &impl Prob
 
         let (node,prod_tp) = match &prod {
             Prod::Prim(p, raw_tp_ref) => (Node::Prim(p.clone()), state.expr.ctx.instantiate(*raw_tp_ref)),
-            Prod::Var(i, tp_ref) => (Node::Var(*i,-1), tp_ref.clone()),
+            Prod::Var(i, tp_ref) => (Node::Var(*i,-1), *tp_ref),
         };
 
         let unnormalized_ll = model.expansion_unnormalized_ll(&node, &state.expr, &hole);
@@ -271,7 +271,7 @@ pub fn add_expansions(state: &mut ThreadState, prods: &[Prod], model: &impl Prob
         }
 
         // unification check
-        if !state.expr.ctx.unify(&hole_tp, &prod_tp.return_type(&state.expr.ctx)).is_ok() {
+        if state.expr.ctx.unify(&hole_tp, &prod_tp.return_type(&state.expr.ctx)).is_err() {
             continue;
         }
         
@@ -410,8 +410,6 @@ pub fn top_down<D: Domain, M: ProbabilisticModel>(
 
 #[derive(Debug, Clone)]
 pub struct WorkItem {
-    tp: Type,
-    env: Vec<Type>,
     tasks: Vec<TaskName>,
     upper_bound: NotNan<f32>,
     lower_bound: NotNan<f32>,
@@ -455,7 +453,7 @@ impl SearchProgress {
         }
         let mut tasks: Vec<TaskName> = self.solutions.keys().cloned().collect();
         tasks.sort();
-        s += &format!("Solutions:");
+        s += "Solutions:";
         for name in &tasks {
             let solns = &self.solutions[name];
             if solns.is_empty() {
@@ -534,8 +532,6 @@ impl Iterator for SearchProgress {
         let single_hole = PartialExpr::single_hole(tp.return_type(&typeset), env.clone(), typeset);
 
         let work_item = WorkItem {
-            tp,
-            env,
             tasks: tasks.clone(),
             upper_bound: self.curr_upper_bound,
             lower_bound: self.curr_upper_bound - self.cfg.step,
@@ -573,6 +569,8 @@ fn search_worker<D: Domain, M: ProbabilisticModel>(shared: Arc<Shared<D,M>>, thr
         let mut new_state = None;
         let mut all_done = true;
 
+        let initial_num_steals = {shared.stats.lock().unwrap().num_steals};
+
         // try work stealing
         for state in shared.thread_states.iter() {
             // LOCK SAFETY: lock will drop at end of for-loop, and we aren't currently holding any locks and will not take
@@ -605,15 +603,16 @@ fn search_worker<D: Domain, M: ProbabilisticModel>(shared: Arc<Shared<D,M>>, thr
                     }
                     // set stolen_from to have just 1 expansion
                     new_state.as_mut().unwrap().save_states[stolen_from].num_expansions = num_to_take;
-
-                    // println!("[Thread {:?}] {}", thread::current().id(), "Stole work".yellow());
-
                     shared.stats.lock().unwrap().num_steals += 1;
 
                     break
                 }
             }
         }
+
+        all_done &= initial_num_steals == {
+            shared.stats.lock().unwrap().num_steals
+        };
 
         if all_done && thread_idx == 0 {
 
@@ -670,7 +669,7 @@ fn search_worker<D: Domain, M: ProbabilisticModel>(shared: Arc<Shared<D,M>>, thr
 
                     let all_solutions = search_progress.solutions.get_mut(&soln.task_name).unwrap();
 
-                    if all_solutions.len() == 1 {
+                    if all_solutions.len() == 0 {
                         // LOCK SAFETY: in no part of the code do we take the search progress lock while already holding the stats lock
                         let mut stats = shared.stats.lock().unwrap();
                         stats.num_never_solved -= 1;
@@ -727,7 +726,6 @@ fn search_in_bounds<D: Domain, M: ProbabilisticModel>(thread_idx: usize, work_it
             }
         }
         
-
         // occasionally transfer stats over. Note num_processed gets reset to 0 during a transfer
         if local_stats.num_processed >= 25_000 {
             shared.stats.lock().unwrap().local.transfer(&mut local_stats);
@@ -772,7 +770,7 @@ fn search_in_bounds<D: Domain, M: ProbabilisticModel>(thread_idx: usize, work_it
             }
             local_stats.num_finished += 1;
 
-            let solved_tasks = check_correctness(&shared, &work_item, &state.expr, &mut local_stats);
+            let solved_tasks = check_correctness(shared, &work_item, &state.expr, &mut local_stats);
 
             for task_name in solved_tasks {
 
@@ -802,7 +800,7 @@ fn search_in_bounds<D: Domain, M: ProbabilisticModel>(thread_idx: usize, work_it
     }
 
     // lock scope
-    { shared.stats.lock().unwrap().local.transfer(&mut local_stats) };
+    shared.stats.lock().unwrap().local.transfer(&mut local_stats);
     // lock scope
     { *shared.thread_states[thread_idx].lock().unwrap() = None; }
     solutions
@@ -853,7 +851,6 @@ fn check_correctness<D: Domain, M: ProbabilisticModel>(shared: &Shared<D,M>, wor
                 }
             }
         }
-        // solved_buf.push((unnormalized_ll, task.name.clone(), expanded.clone()));
         if solved {
             solved_tasks.push(task.name.clone());
         }
